Rename namespace: Files => Templates
