+++ Extdn/Samples/Templates/objectmanager.phtml
